requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.26.3
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Cython==0.29.33
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.89.1
ffmpeg-python==0.2.0
filelock==3.9.0
Flask==2.2.2
Flask-Cors==3.0.10
frozenlist==1.3.3
future==0.18.3
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-vision==3.3.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.12.0
idna==3.4
importlib-metadata==6.0.0
isodate==0.6.1
itsdangerous==2.1.2
jaraco.context==4.3.0
Jinja2==3.1.2
MarkupSafe==2.1.2
more-itertools==9.0.0
msrest==0.7.1
multidict==6.0.4
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
openai==0.26.4
openai-whisper==20230124
opencv-python==4.7.0.68
orjson==3.8.5
os-tools==4.56
packaging==23.0
Pillow==9.4.0
pkg_resources==0.0.0
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pydantic==1.10.4
pymongo==4.3.3
pytesseract==0.3.10
python-dotenv==0.21.1
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tesseract==0.1.3
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.1
tqdm==4.64.1
transformers==4.26.0
typing_extensions==4.4.0
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
Werkzeug==2.2.2
whisper==1.1.10
wolframalpha==5.0.0
wolframclient==1.1.7
xmltodict==0.13.0
yarl==1.8.2
zipp==3.12.0
